+++ src/Subscription/ChangeLog.md
@@ -19,8 +19,8 @@
 -->
 ## Upcoming Release
 * Removed the default value for the `New-AzSubscriptionAlias` cmdlet parameter `SubscriptionId`.
-* Removed the default value for the `Get-AzSubscriptionAcceptOwnershipStatus` cmdlet parameter `SubscriptionId` and mark the parameter as required.
-* Removed the default value for the `Invoke-AzSubscriptionAcceptOwnership` cmdlet parameter `SubscriptionId` and mark the parameter as required.
+* Removed the default value for the `Get-AzSubscriptionAcceptOwnershipStatus` cmdlet parameter `SubscriptionId` and mark the parameter as mandatory.
+* Removed the default value for the `Invoke-AzSubscriptionAcceptOwnership` cmdlet parameter `SubscriptionId` and mark the parameter as mandatory.
 
 ## Version 0.10.0
 * Upgrade API version to 2021-10-01.